requirements.txt
--index-url https://download.pytorch.org/whl/cu118

torch
torchvision
torchaudio

diffusers
transformers
accelerate

regex
fastapi
uvicorn
psycopg2-binary
pydantic
crewai
langchain
openai

python-dotenv
gradio